Update temperature to float
The temperature is stored in struct aht20_sensor as a float.
README and example is updated to reflect the update.

+++ source/AHT20.h
@@ -49,7 +49,7 @@ struct aht20_sensor
 {
     char  buf[SENSOR_DATA_LEN];
     float humidity;
-    int   temperature;
+    float   temperature;
     int   lg_I2C_handle;
     int   I2C_dev;
 };
